Compatibility: Need to use old regexp engine for correct capture group assignment
t2000-YankMatches.vim and t2100-YankUniqueMatches.vim fail with Vim 7.4.1689 because the NFA-based regexp engine somehow captures a single "a" replacement to the second group instead of the first (optional) one. Changing the second group's .* to .\{-} does not help. The problem is gone in Vim 8.2.346.
Switch to the old regexp engine, but only for the capturing; for the pure matching the assignment to groups does not matter. Inside ingo#cmdargs#substitute#Parse(), we don't have a good way to influence the matching, unfortunately, but apparently /{pat}/{repl}/{reg} matching isn't affected by the bug (because a:options.flagsExpr is part of an even larger regexp).